p112, n8vem-mark4: Provide boot device parsing method
