+++ Kernel/platform-n8vem-mark4/README
@@ -27,7 +27,7 @@ load.
 The file "Kernel/platform-n8vem-mark4/diskboot.bin" is an UNA BIOS compatible
 boot sector which will allow the system to be booted directly from an IDE or SD
 drive. The boot sector will load the kernel image starting from the third
-sector of the drive.
+sector of the drive (ie counting from 0, sector number 2).
 
 The kernel image (fuzix.bin) should be written to the disk drive from the third
 sector onwards, in the space before the first partition starts. Be careful to
@@ -47,6 +47,20 @@ table, you can use dd as follows;
 To write the kernel image to the disk drive, use dd as follows;
     dd if=fuzix.bin bs=512 seek=2 of=/dev/sdX
 
+The root filesystem device can be specified as either a decimal number, or as a
+a device name (without the /dev prefix); hda, hdb, hda1, hda2, etc.
+
 When booting the system from both UNA BIOS and CP/M you can specify the root
-filesystem device number on the command line after the command name or unit
-number.
+filesystem device on the command line after the command name or unit number.
+
+For example, with root filesystem on /dev/hdb1 (minor #17) at the CP/M command
+prompt:
+
+    A> FUZIX hdb1
+or  A> FUZIX 17
+
+And to boot from unit 3 in UNA BIOS:
+
+    Boot UNA unit number or ROM? [R,X,0..3] (R): 3 hdb1
+or  Boot UNA unit number or ROM? [R,X,0..3] (R): 3 17
+
