+++ Kernel/platform-n8vem-mark4/main.c
@@ -24,19 +24,6 @@ void z180_timer_interrupt(void)
     timer_interrupt();
 }
 
-
-void pagemap_init(void)
-{
-    int i;
-
-    /* N8VEM SBC Mark IV has RAM in the top 512K of physical memory. 
-     * First 64K is used by the kernel. 
-     * Each process gets the full 64K for now.
-     * Page size is 4KB. */
-    for(i = 0x90; i < 0x100; i+=0x10)
-        pagemap_add(i);
-}
-
 void platform_idle(void)
 {
     /* Let's go to sleep while we wait for something to interrupt us;
@@ -61,12 +48,4 @@ void platform_interrupt(void)
         default:
             return;
     }
-}
-
-void map_init(void)
-{
-    /* clone udata and stack into a regular process bank, return with common memory
-       for the new process loaded */
-    copy_and_map_process(&init_process->p_page);
-    /* kernel bank udata (0x300 bytes) is never used again -- could be reused? */
 }
