+++ Kernel/platform-p112/fuzix.lnk
@@ -31,6 +31,7 @@ mm.rel
 bankfixed.rel
 swap.rel
 devsys.rel
+platform-p112/discard.rel
 platform-p112/devtty.rel
 platform-p112/devide.rel
 platform-p112/devide_discard.rel
