+++ Kernel/platform-p112/main.c
@@ -26,20 +26,6 @@ void z180_timer_interrupt(void)
     timer_interrupt();
 }
 
-
-void pagemap_init(void)
-{
-    int i;
-
-    /* P112 has RAM across the full physical 1MB address space
-     * First 64K is used by the kernel. 
-     * Each process gets the full 64K for now.
-     * Page size is 4KB. */
-    for(i = 0x10; i < 0x100; i+=0x10){
-        pagemap_add(i);
-    }
-}
-
 void platform_idle(void)
 {
     /* Let's go to sleep while we wait for something to interrupt us;
@@ -68,12 +54,4 @@ void platform_interrupt(void)
             kprintf("[int:%d?]", irqvector);
             return;
     }
-}
-
-void map_init(void)
-{
-    /* clone udata and stack into a regular process bank, return with common memory
-       for the new process loaded */
-    copy_and_map_process(&init_process->p_page);
-    /* kernel bank udata (0x300 bytes) is never used again -- could be reused? */
 }
